requirements.txt
jinja2
python-multipart
aiofiles
openai
cryptography==44.0.2
fastapi==0.115.8
httpx==0.28.1
PyJWT==2.10.1
python-dotenv==1.0.1
uvicorn==0.34.0
requests==2.32.3